Create niceprint and UPLDRConstants class
- for now: branches blacklist for TravisCI testing
